+++ assets/js/Containers/FrevaGPT/components/Snippets/BotLoadingSpinner.js
@@ -10,7 +10,7 @@ const BotLoadingSpinner = forwardRef(({ dynamicAnswer, loading }, ref) => {
       {loading && !dynamicAnswer ? (
         <Row className="mb-3">
           <Col md={3}>
-            <Card className="shadow-sm card-body border-0 border-bottom mb-3 bg-light d-flex flex-row align-items-center">
+            <Card className="bot-shadow br-8 card-body border-0 border-bottom mb-3 bg-light d-flex flex-row align-items-center">
               <Spinner size="sm" />
               <span className="ms-2">
                 {ref.lastVariant.current === "Code"