+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
+## [v2603.0.0]
+### Added
+- Handling of reference links in bot answers
+### Changed
+- Improved styling of bot
+- Improved performance by switching to highlight.js (dropping react-syntax-highlighter)
+### Fixed
+- Copying code prodiuced by chatbot
 ## [v2602.0.0]
 ### Added
 - Zarr aggregation feature